Fix `+nnn` bug in FromString
While doing experiments with the dev.fuzz branch of the Go tool chain,
I've found a bug in FromString() within 1 second of fuzz time:

Because I don't use a regexp in the parseIdentifier, but instead simply
use strconv.Atoi, it is possible to sneak in a sign symbol.  `-` is
detected, but + is not.  There is not much harm, as an AIRAC identifier
like "+911" is interpreted as "0911", but technically it is not a valid
identifier and should be rejected.  Also, although a `-` sign correctly
yields errors, it is not clear in the code, why.

This change rejects + and - signs explicitly.